Move test_lambdas to new style
